Add explicit int cast for rotating log max_files values

+++ config/logging.php
@@ -69,7 +69,7 @@ return [
             'driver' => 'daily',
             'path' => storage_path('logs/laravel.log'),
             'level' => env('LOG_LEVEL', 'debug'),
-            'max_files' => env('LOG_DAILY_DAYS', 14),
+            'max_files' => (int) env('LOG_DAILY_DAYS', 14),
             'replace_placeholders' => true,
         ],
 
@@ -77,7 +77,7 @@ return [
             'driver' => 'monthly',
             'path' => storage_path('logs/laravel.log'),
             'level' => env('LOG_LEVEL', 'debug'),
-            'max_files' => env('LOG_MONTHLY_MONTHS', 3),
+            'max_files' => (int) env('LOG_MONTHLY_MONTHS', 3),
             'replace_placeholders' => true,
         ],
